requirements.txt
pandas==2.2.2
pymysql==1.1.1
python-dotenv==1.0.1
google-cloud-storage==2.16.0
mysql-connector-python==8.4.0
apache-airflow-providers-mysql==5.6.1
firebase-admin==6.5.0